Daemonize
Uses a second level child process to run as a true daemon instead of
relying on the sigaction based handler. The latter process broke
subsequent calls by other components of OpenVPN which prevented it's use
in pfSense.

This approach was based on the standard duo_openvpn c wrapper found
here:
https://github.com/duosecurity/duo_openvpn/blob/master/duo_openvpn.c